backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
openai==1.58.1
beautifulsoup4==4.12.3
pydantic-settings==2.7.1
python-multipart==0.0.20
lxml==5.3.0
